+++ libs/secret_text_box/secret_text_box.h
@@ -31,7 +31,7 @@ private:
 public:
 	SecretTextBox(){}
 
-	SecretTextBox(int max_len, const std::string& font_path, int font_size,
+	SecretTextBox(size_t max_len, const std::string& font_path, int font_size,
 			Window& window);
 
 	/**
